+++ tests/routes/test_bring_beer_routes.py
@@ -50,7 +50,7 @@ def test_read_wrong_bring_beer_id(client_fixture):
 
     response = client_fixture.get(f"/bringbeer/{wrong_id}")
     assert response.status_code == 404
-    assert response.json()["detail"] == f"Bring beer with id '{wrong_id}' not found!"
+    assert response.json()["detail"] == f"BRING_BEER with id '{wrong_id}' not found!"
 
 
 def test_delete_bring_beer(client_fixture, get_admin_token):
@@ -82,7 +82,7 @@ def test_delete_wrong_bring_beer(client_fixture, get_admin_token):  #
         f"/bringbeer/{wrong_id}", headers={"Authorization": f"Bearer {get_admin_token}"}
     )
     assert response.status_code == 404
-    assert response.json()["detail"] == f"Bring beer with id '{wrong_id}' not found!"
+    assert response.json()["detail"] == f"BRING_BEER with id '{wrong_id}' not found!"
 
 
 def test_delete_bring_beer_invalid_token(client_fixture, get_invalid_token):
@@ -99,7 +99,7 @@ def test_delete_bring_beer_invalid_token(client_fixture, get_invalid_token):
         headers={"Authorization": f"Bearer {get_invalid_token}"},
     )
     assert response.status_code == 401
-    assert response.json()["detail"] == "Invalid token or role"
+    assert response.json()["detail"] == "Invalid token"
 
 
 def test_update_bring_beer_name(client_fixture):
@@ -128,7 +128,7 @@ def test_update_wrong_bring_beer(client_fixture):
 
     response = client_fixture.patch(f"/bringbeer/{wrong_id}", json={})
     assert response.status_code == 404
-    assert response.json()["detail"] == f"Bring beer with id '{wrong_id}' not found!"
+    assert response.json()["detail"] == f"BRING_BEER with id '{wrong_id}' not found!"
 
 
 def test_read_done_bring_beer(client_fixture):