+++ tests/test_auth.py
@@ -8,8 +8,9 @@ from fastapi import HTTPException
 import pytest
 from sqlmodel import Session
 
-from auth.auth_methods import auth_is_user, auth_is_admin
+from auth.auth_methods import auth_is_user, auth_is_admin, auth_is_user_or_admin
 from auth.login_routes import authenticate_user
+from exceptions import InvalidUserException, InvalidTokenException, InvalidRoleException
 from tests.helper_methods import create_user, create_team
 
 
@@ -117,7 +118,7 @@ def test_auth_is_user(get_user_token):
     :param get_user_token: Test user token.
     :return: None
     """
-    assert auth_is_user(1, get_user_token)
+    assert not auth_is_user(1, get_user_token)
 
 
 def test_auth_is_not_user(get_user_token):
@@ -126,7 +127,8 @@ def test_auth_is_not_user(get_user_token):
     :param get_user_token: Test user token.
     :return: None
     """
-    assert not auth_is_user(15, get_user_token)
+    with pytest.raises(InvalidUserException):
+        auth_is_user(15, get_user_token)
 
 
 def test_auth_user_with_invalid_token(get_invalid_token):
@@ -135,7 +137,8 @@ def test_auth_user_with_invalid_token(get_invalid_token):
     :param get_invalid_token: Test invalid token.
     :return: None
     """
-    assert not auth_is_user(15, get_invalid_token)
+    with pytest.raises(InvalidTokenException):
+        auth_is_user(15, get_invalid_token)
 
 
 def test_auth_is_admin(get_admin_token):
@@ -144,7 +147,7 @@ def test_auth_is_admin(get_admin_token):
     :param get_admin_token: Test admin token.
     :return: None
     """
-    assert auth_is_admin(get_admin_token)
+    assert not auth_is_admin(get_admin_token)
 
 
 def test_auth_is_not_admin(get_user_token):
@@ -153,7 +156,8 @@ def test_auth_is_not_admin(get_user_token):
     :param get_user_token: Test user token.
     :return: None
     """
-    assert not auth_is_admin(get_user_token)
+    with pytest.raises(InvalidRoleException):
+        auth_is_admin(get_user_token)
 
 
 def test_auth_admin_with_invalid_token(get_invalid_token):
@@ -162,4 +166,15 @@ def test_auth_admin_with_invalid_token(get_invalid_token):
     :param get_invalid_token: Test invalid token.
     :return: None
     """
-    assert not auth_is_admin(get_invalid_token)
+    with pytest.raises(InvalidTokenException):
+        auth_is_admin(get_invalid_token)
+
+
+def test_is_user_or_admin_invalid_token(get_invalid_token):
+    """
+    Test is_user_or_admin with an invalid token exception.
+    :param get_invalid_token: Test invalid token.
+    :return: None
+    """
+    with pytest.raises(InvalidTokenException):
+        auth_is_user_or_admin(12, get_invalid_token)